refact: #49 Combine TodayQuiz's ViewController
