+++ GaNaDa/GaNaDa/View/TodayQuiz/TodayQuizCollectionVC.swift
@@ -11,6 +11,10 @@ import UIKit
 final class TodayQuizViewController: UIViewController {
 
     @IBOutlet weak var todayQuizCollectionView: UICollectionView!
+    @IBOutlet weak var userImage: UIImageView!
+    @IBOutlet weak var userLevel: UILabel!
+    @IBOutlet weak var userName: UILabel!
+    @IBOutlet weak var userExp: UIProgressView!
     
     var todayQuizs: [Quiz] = [Quiz(question: "나는 ios 개발자가 * 싶다.", type: QuizType.blank, rightAnswer: "되고", wrongAnswer: "돼고"),
                               Quiz(question: "정말 너를 * 좋니.", type: QuizType.blank, rightAnswer: "어떡하면", wrongAnswer: "어떻하면"),
@@ -19,7 +23,12 @@ final class TodayQuizViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+        saveUserData(userName: "박가네감자탕둘째며느리의셋째아들")
+        requestUserData()
+        
+        userExp.layer.sublayers![1].cornerRadius = 6
+        userExp.subviews[1].clipsToBounds = true
+        
         let todayQuizBlankCellNib = UINib(nibName: "QuizTypeBlank", bundle: nil)
         
         todayQuizCollectionView.register(todayQuizBlankCellNib, forCellWithReuseIdentifier: "todayQuizBlankCell")
@@ -28,6 +37,19 @@ final class TodayQuizViewController: UIViewController {
         todayQuizCollectionView.dataSource = self
         
         todayQuizCollectionView.collectionViewLayout = creatCompositionalLayout()
+    }
+    
+    func saveUserData(userName: String) {
+        if UserDefaults.standard.object(forKey: "userName") == nil {
+            UserDefaults.standard.setValue(userName, forKey: "userName")
+            UserDefaults.standard.setValue(280, forKey: "userExp")
+        }
+    }
+    
+    func requestUserData() {
+        userLevel.text = level(rawValue: UserDefaults.standard.integer(forKey: "userExp") / 100)?.name
+        userName.text = UserDefaults.standard.string(forKey: "userName")
+        userExp.progress = Float(UserDefaults.standard.integer(forKey: "userExp") % 100) / 100.0
     }
 }
 
@@ -73,5 +95,67 @@ extension TodayQuizViewController: UICollectionViewDataSource{
         cell.data = self.todayQuizs[indexPath.item]
         cell.quizIndex.text = "문제 \(indexPath.item + 1)"
         return cell
+    }
+}
+
+enum level: Int {
+    case lowNine = 0
+    case highNine
+    case lowEight
+    case highEight
+    case lowSeven
+    case highSeven
+    case lowSix
+    case highSix
+    case lowFive
+    case highFive
+    case lowFour
+    case highFour
+    case lowThree
+    case highThree
+    case lowTwo
+    case highTwo
+    case lowOne
+    case highOne
+    
+    var name: String {
+        switch self {
+        case .lowNine:
+            return "종 9품"
+        case .highNine:
+            return "정 9품"
+        case .lowEight:
+            return "종 8품"
+        case .highEight:
+            return "정 8품"
+        case .lowSeven:
+            return "종 7품"
+        case .highSeven:
+            return "정 7품"
+        case .lowSix:
+            return "종 6품"
+        case .highSix:
+            return "정 6품"
+        case .lowFive:
+            return "종 5품"
+        case .highFive:
+            return "정 5품"
+        case .lowFour:
+            return "종 4품"
+        case .highFour:
+            return "정 4품"
+        case .lowThree:
+            return "종 3품"
+        case .highThree:
+            return "정 3품"
+        case .lowTwo:
+            return "종 2품"
+        case .highTwo:
+            return "정 2품"
+        case .lowOne:
+            return "종 1품"
+        case .highOne:
+            return "정 1품"
+        }
     }
 }
